+++ internal/outwriter/output_metrics.go
@@ -10,59 +10,6 @@ import (
 	"github.com/huangsam/hotspot/schema"
 )
 
-// getDisplayNameForMode returns the display name with emoji for a given mode name.
-func getDisplayNameForMode(modeName string) string {
-	switch modeName {
-	case "hot":
-		return "🔥 HOT"
-	case "risk":
-		return "⚠️  RISK"
-	case "complexity":
-		return "🧩 COMPLEXITY"
-	case "stale":
-		return "🕰️  STALE"
-	default:
-		return strings.ToUpper(modeName)
-	}
-}
-
-// getDisplayWeightsForMode returns the weights to display for a given scoring mode.
-// Uses active weights if available, otherwise falls back to defaults.
-func getDisplayWeightsForMode(mode schema.ScoringMode, activeWeights map[schema.ScoringMode]map[schema.BreakdownKey]float64) map[string]float64 {
-	// Start with default weights
-	defaultWeights := schema.GetDefaultWeights(mode)
-
-	// Convert BreakdownKey map to string map for backward compatibility
-	weights := make(map[string]float64)
-	for k, v := range defaultWeights {
-		weights[string(k)] = v
-	}
-
-	// Override with active weights if provided
-	if activeWeights != nil {
-		if modeWeights, ok := activeWeights[mode]; ok {
-			// Only override weights that are actually customized
-			for k, v := range modeWeights {
-				weights[string(k)] = v
-			}
-		}
-	}
-
-	return weights
-}
-
-// formatWeights formats weights for display in formulas.
-func formatWeights(weights map[string]float64, factorKeys []string) string {
-	var parts []string
-	for _, key := range factorKeys {
-		if weight, ok := weights[key]; ok && weight > 0 {
-			factorName := strings.ToLower(strings.TrimPrefix(key, "breakdown_"))
-			parts = append(parts, fmt.Sprintf("%.2f*%s", weight, factorName))
-		}
-	}
-	return strings.Join(parts, "+")
-}
-
 // WriteMetricsDefinitions displays the formal definitions of all scoring modes.
 // This is a static display that does not require Git analysis.
 func WriteMetricsDefinitions(activeWeights map[schema.ScoringMode]map[schema.BreakdownKey]float64, cfg *contract.Config) error {
@@ -143,58 +90,6 @@ func writeMetricsCSV(renderModel *schema.MetricsRenderModel, cfg *contract.Confi
 		defer writer.Flush()
 		return writeCSVMetrics(writer, renderModel)
 	}, "Wrote CSV")
-}
-
-// buildMetricsRenderModel constructs the complete render model with all processed data.
-func buildMetricsRenderModel(activeWeights map[schema.ScoringMode]map[schema.BreakdownKey]float64) *schema.MetricsRenderModel {
-	modes := []schema.MetricsMode{
-		{
-			Name:       "hot",
-			Purpose:    "Activity hotspots - high recent activity & volatility",
-			Factors:    []string{"Commits", "Churn", "Contributors", "Age", "Size"},
-			FactorKeys: []string{string(schema.BreakdownCommits), string(schema.BreakdownChurn), string(schema.BreakdownContrib), string(schema.BreakdownAge), string(schema.BreakdownSize)},
-		},
-		{
-			Name:       "risk",
-			Purpose:    "Knowledge risk/bus factor - concentrated ownership",
-			Factors:    []string{"InvContributors", "Gini", "Age", "Churn", "Commits", "LOC", "Size"},
-			FactorKeys: []string{string(schema.BreakdownInvContrib), string(schema.BreakdownGini), string(schema.BreakdownAge), string(schema.BreakdownChurn), string(schema.BreakdownCommits), string(schema.BreakdownLOC), string(schema.BreakdownSize)},
-		},
-		{
-			Name:       "complexity",
-			Purpose:    "Technical debt - large, old files with high maintenance burden",
-			Factors:    []string{"Age", "Churn", "Commits", "LOC", "LowRecent", "Size"},
-			FactorKeys: []string{string(schema.BreakdownAge), string(schema.BreakdownChurn), string(schema.BreakdownCommits), string(schema.BreakdownLOC), string(schema.BreakdownLowRecent), string(schema.BreakdownSize)},
-		},
-		{
-			Name:       "stale",
-			Purpose:    "Maintenance debt - important files untouched recently",
-			Factors:    []string{"InvRecent", "Age", "Size", "Commits", "Contributors"},
-			FactorKeys: []string{string(schema.BreakdownInvRecent), string(schema.BreakdownAge), string(schema.BreakdownSize), string(schema.BreakdownCommits), string(schema.BreakdownContrib)},
-		},
-	}
-	modesWithData := make([]schema.MetricsModeWithData, len(modes))
-
-	for i, mode := range modes {
-		weights := getDisplayWeightsForMode(schema.ScoringMode(mode.Name), activeWeights)
-		formula := formatWeights(weights, mode.FactorKeys)
-
-		modesWithData[i] = schema.MetricsModeWithData{
-			MetricsMode: mode,
-			Weights:     weights,
-			Formula:     formula,
-		}
-	}
-
-	return &schema.MetricsRenderModel{
-		Title:       "Hotspot Scoring Modes",
-		Description: "All scores = weighted sum of normalized factors",
-		Modes:       modesWithData,
-		SpecialRelationship: map[string]string{
-			"description": "RISK Score = HOT Score / Ownership Diversity Factor",
-			"note":        "(Factor ↓ when few contributors → RISK Score ↑)",
-		},
-	}
 }
 
 // writeJSONMetrics writes the metrics definitions in JSON format.
